Update default.rb
Change the repo location for the vddk-libs to the enterprise-devel location. The old URL was broken. Not sure if we want to put more logic there to detect the euca version requested and add that into the url or not. But, this fix will pull from the latest devel packages for the vddk-libs, which presumably doesn't change often