Move Set blockstoragemanager block outside the ruby block.
This follows the same logic, but the problem of notifiying
an execute block and interpolating the #{cluster} variable
is no longer present, allowing each cluster to be configured
successfully.  The logic is not changed because if the Timeout
is reached it raises an error and exits, but if not we continue
on to set the blockstoragemanager when ready.