Check for broken objectstorage state; use $IP.xip.io in faststart.ini
We will now check for both enabled and broken state while waiting
for objectstorage service.  But for compute/cloudformation we still
want to only wait for enabled state.

Also use $IP.xip.io as the region for faststart.ini and in the turorial
vs. localhost which was conflicting with the euca2ools localhost.ini
so now we should avoid situations where users are reporting errors
in create-first-resources.rb.